pyproject.toml
[tool.poetry]
name = "monetary-movements-test"
version = "0.1.0"
description = ""
authors = ["macadden <mati.macadden@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"


[tool.poetry.group.dev.dependencies]
django = "4.2.7"
djangorestframework = "3.14.0"
django-cors-headers = "4.3.1"
typing-extensions = "4.9.0"
djangorestframework-simplejwt = "5.3.1"
django-simple-history = "3.4.0"
psycopg2-binary = "2.9.9"
django-filter = "23.4"
requests = "2.31.0"
requests-mock = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
